# backend/src/config/mobileOpenApi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lidifin Mobile API",
    "version": "1.0.0",
    "description": "Dedicated mobile-facing API contract for native clients. All paths are relative to /api. Primary authentication is username/password via JWT access + refresh tokens."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Relative API base path"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Primary mobile authentication. Obtain with POST /auth/login and refresh with POST /auth/refresh."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Secondary device authentication for paired devices and integrations."
      },
      "streamTokenAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "Optional query token for stream URLs when attaching Authorization headers is inconvenient."
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Web UI session authentication. Included for completeness."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" },
          "message": { "type": "string" },
          "details": {}
        }
      },
...
      "ApiKey": {
        "type": "object",
        "required": ["id", "name", "createdAt"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "lastUsed": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
    }
  },
  "paths": {
...
    "/device-link/generate": {
      "post": {
        "tags": ["Device Link"],
        "summary": "Generate a short-lived device link code",
        "description": "Optional secondary pairing flow for clients that should link to an already authenticated account.",
        "security": [
          { "bearerAuth": [] },
          { "sessionAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "Link code created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": { "type": "string", "example": "AB12CD" },
                    "expiresAt": { "type": "string", "format": "date-time" },
                    "expiresIn": { "type": "integer", "example": 300 }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/device-link/verify": {
      "post": {
        "tags": ["Device Link"],
        "summary": "Exchange a device link code for an API key",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["code"],
                "properties": {
                  "code": { "type": "string", "example": "AB12CD" },
                  "deviceName": { "type": "string", "example": "Pixel 9" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device linked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "apiKey": { "type": "string" },
                    "userId": { "type": "string" },
                    "username": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired code",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Code not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/device-link/status/{code}": {
      "get": {
        "tags": ["Device Link"],
        "summary": "Poll the status of a device link code",
        "parameters": [
          {
            "in": "path",
            "name": "code",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Current code status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": ["pending", "used", "expired"]
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "usedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "deviceName": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Code not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/device-link/devices": {
      "get": {
        "tags": ["Device Link"],
        "summary": "List linked devices",
        "security": [
          { "bearerAuth": [] },
          { "sessionAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "Linked devices returned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ApiKey" }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/device-link/devices/{id}": {
      "delete": {
        "tags": ["Device Link"],
        "summary": "Revoke a linked device",
        "security": [
          { "bearerAuth": [] },
          { "sessionAuth": [] }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Device revoked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Device not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  }
}